refactor(website): inline homepage sections, delete orphaned home components
The homepage renders only 4 sections after the #1255 content-first redesign.
Inline the three homepage-only sections (hero, pricing, CTA) into a single
apps/website/app/(public)/(home)/home-content.tsx so all homepage copy lives in
one file, easy to review top-to-bottom. _footer stays a shared component — it is
imported site-wide by PageLayout and cannot be inlined.

Delete 23 orphaned home/_*.tsx components (+ their .spec.tsx and
_features.helpers.ts), leftovers from the pre-#1255 long-scroll design. Verified
none are imported by any of the 24 public route pages (_testimonials and _formats
were also dead despite the original ask assuming otherwise). Replace the three
deleted section specs with one consolidated home-content.spec.tsx.

Green-field cleanup per CLAUDE.md: ~4.3k lines of dead code removed.
