--- apps/website/packages/components/home/_features.helpers.ts
@@ -1,77 +1,0 @@
-import {
-  HiChartBar,
-  HiChatBubbleLeftRight,
-  HiCpuChip,
-  HiNewspaper,
-  HiPhoto,
-  HiShare,
-  HiSparkles,
-  HiUserCircle,
-  HiVideoCamera,
-} from 'react-icons/hi2';
-
-export const allFeatures = [
-  {
-    color: 'bg-purple-500',
-    description:
-      'Generate AI-powered digital twins with realistic voices and personas for authentic content.',
-    icon: HiUserCircle,
-    label: 'Avatar Twins',
-  },
-  {
-    color: 'bg-pink-500',
-    description:
-      'Create professional videos from text, images, or prompts. AI handles editing, voiceovers, and effects.',
-    icon: HiVideoCamera,
-    label: 'Video Generation',
-  },
-  {
-    color: 'bg-blue-500',
-    description:
-      'Produce high-quality images for any use case. From product shots to social media graphics.',
-    icon: HiPhoto,
-    label: 'Image Generation',
-  },
-  {
-    color: 'bg-orange-500',
-    description:
-      'Generate engaging, SEO-optimized articles at scale. AI writes, formats, and publishes automatically.',
-    icon: HiNewspaper,
-    label: 'Article Writing',
-  },
-  {
-    color: 'bg-green-500',
-    description:
-      'Access Genfeed directly from ChatGPT. Create and publish content without leaving your chat.',
-    icon: HiChatBubbleLeftRight,
-    label: 'ChatGPT Integration',
-  },
-  {
-    color: 'bg-indigo-500',
-    description:
-      'Connect with any AI agent using Model Context Protocol. Seamlessly integrate with your workflow.',
-    icon: HiCpuChip,
-    label: 'MCP Integration',
-  },
-  {
-    color: 'bg-cyan-500',
-    description:
-      'Post directly to Twitter, Instagram, LinkedIn, and more. Schedule and automate your entire pipeline.',
-    icon: HiShare,
-    label: 'Social Publishing',
-  },
-  {
-    color: 'bg-yellow-500',
-    description:
-      'Build custom workflows. From research to publishing, let AI handle the repetitive tasks.',
-    icon: HiSparkles,
-    label: 'AI Workflow Automation',
-  },
-  {
-    color: 'bg-emerald-500',
-    description:
-      'Track engagement, reach, and ROI across all channels. Data-driven insights for better content.',
-    icon: HiChartBar,
-    label: 'Performance Analytics',
-  },
-];
