--- apps/website/packages/components/home/_open-source.spec.tsx
@@ -1,9 +1,0 @@
-import * as Module from '@web-components/home/_open-source';
-import { describe, expect, it } from 'vitest';
-
-describe('OpenSource Component', () => {
-  it('exports a default component', () => {
-    expect(Module).toHaveProperty('default');
-    expect(typeof Module.default).toBe('function');
-  });
-});
