--- apps/website/packages/components/home/_pricing.spec.tsx
@@ -1,9 +1,0 @@
-import * as Module from '@web-components/home/_pricing';
-import { describe, expect, it } from 'vitest';
-
-describe('Pricing Component', () => {
-  it('exports a default component', () => {
-    expect(Module).toHaveProperty('default');
-    expect(typeof Module.default).toBe('function');
-  });
-});
