--- apps/website/packages/components/home/_tool-stack.spec.tsx
@@ -1,9 +1,0 @@
-import * as Module from '@web-components/home/_tool-stack';
-import { describe, expect, it } from 'vitest';
-
-describe('ToolStack Component', () => {
-  it('exports a default component', () => {
-    expect(Module).toHaveProperty('default');
-    expect(typeof Module.default).toBe('function');
-  });
-});
