--- apps/website/packages/components/home/_top-brands.spec.tsx
@@ -1,9 +1,0 @@
-import * as Module from '@web-components/home/_top-brands';
-import { describe, expect, it } from 'vitest';
-
-describe('TopBrands Component', () => {
-  it('exports a default component', () => {
-    expect(Module).toHaveProperty('default');
-    expect(typeof Module.default).toBe('function');
-  });
-});
